send entire dataset to DownloadDatasetSummary
